suivi_recherche/requirements.txt
Django
djangorestframework
mysqlclient
django-cors-headers
drf-yasg
requests
setuptools
django-debug-toolbar
python-dotenv
django-filter